# docs/openapi.yaml
openapi: 3.1.0
info:
  title: CityConnect API
  version: "1.0.0"
  description: |
    Municipal CRM and service-request API.

    Three kinds of caller reach this API, and they authenticate differently:

    - **Staff** hold an HTTP-only session cookie, established by signing in
      through C2 (`/auth/login` → C2 → `/auth/callback`). There is no local
      account path.
    - **Connected systems** present a `cc_pat_…` personal access token as a
      bearer credential. A token can only ever narrow its owner's authority.
    - **C2 itself** calls two endpoints server-to-server: the Service Card
      callout (`/citizens/{sub}/status`, authenticated by a short-lived RS256
      assertion) and back-channel logout (a signed `logout_token`).

    Errors are `application/problem+json` and carry a stable `code` field.
    Switch on that rather than on the prose.
  contact:
    name: CityConnect
    url: https://github.com/jjamieson1/CityConnect
servers:
  - url: https://city.example.gov/cityconnect/api
    description: Production
  - url: http://localhost:4021/api
    description: Local development
...
security:
  - sessionCookie: []
  - bearerToken: []
paths:
...
  /requests:
    get:
      tags: [Requests]
      summary: List service requests
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: string }, description: "Comma-separated" }
        - { name: priority, in: query, schema: { type: string } }
        - { name: queueId, in: query, schema: { type: string } }
        - { name: departmentId, in: query, schema: { type: string } }
        - { name: assigneeUserId, in: query, schema: { type: string } }
        - { name: contactId, in: query, schema: { type: string } }
        - { name: openOnly, in: query, schema: { type: boolean } }
        - { name: unassigned, in: query, schema: { type: boolean } }
        - { name: breached, in: query, schema: { type: boolean } }
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - $ref: "#/components/parameters/Sort"
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Page"
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: "#/components/schemas/ServiceRequest" }

    post:
      tags: [Requests]
      summary: Create a service request
      description: |
        Validates the service type's intake form, applies routing rules,
        computes SLA targets from a business calendar, and queues the
        acknowledgement notification.

        Partner clients should send an `Idempotency-Key`: integrations retry,
        and a duplicate request is a duplicate work order for a crew.
      parameters:
        - { name: Idempotency-Key, in: header, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateRequest" }
      responses:
        "201":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ServiceRequest" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /requests/{id}:
    parameters: [{ $ref: "#/components/parameters/Id" }]
    get:
      tags: [Requests]
      summary: Get a request
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ServiceRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [Requests]
      summary: Update a request
      description: |
        Send the `version` you read. A mismatch returns `409 stale_version`
        rather than overwriting — two agents on one ticket is routine.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                subject: { type: string }
                description: { type: string }
                priority: { $ref: "#/components/schemas/Priority" }
                version: { type: integer }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ServiceRequest" }
        "409": { $ref: "#/components/responses/Conflict" }
  /requests/{id}/transition:
    parameters: [{ $ref: "#/components/parameters/Id" }]
    post:
      tags: [Requests]
      summary: Change status
      description: |
        Moves the request through the state machine. Illegal moves are rejected
        with `409 bad_transition`, never written — an unconstrained status
        column makes every report about cycle time meaningless.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [to]
              properties:
                to: { $ref: "#/components/schemas/RequestStatus" }
                note: { type: string }
                resolutionCode: { type: string }
                notifyCitizen: { type: boolean }
                version: { type: integer }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ServiceRequest" }
        "409": { $ref: "#/components/responses/Conflict" }
  /requests/{id}/comments:
    parameters: [{ $ref: "#/components/parameters/Id" }]
    post:
      tags: [Requests]
      summary: Add a comment
      description: |
        A `citizen`-visible comment is what the Service Card renders as "what
        has been done" and what a notification quotes. Visibility is an
        explicit choice; the default is internal.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                body: { type: string }
                visibility: { type: string, enum: [internal, citizen] }
                notifyCitizen: { type: boolean }
      responses:
        "201":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Comment" }
  /requests/bulk:
    post:
      tags: [Requests]
      summary: Apply an action to many requests
      description: Returns `207` when any item failed, with per-item outcomes.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [requestIds, operation]
              properties:
                requestIds: { type: array, items: { type: string }, maxItems: 500 }
                operation: { type: string, enum: [assign, transition, priority, queue, tag] }
      responses:
        "200": { description: Every item succeeded }
        "207":
          description: Partial success
          content:
            application/json:
              schema:
                type: object
                properties:
                  succeeded: { type: array, items: { type: string } }
                  failed: { type: object, additionalProperties: { type: string } }

  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: cc_session
      description: Set by the C2 sign-in callback. HTTP-only, SameSite=Lax.
    bearerToken:
      type: http
      scheme: bearer
      description: A `cc_pat_…` personal access token, for connected systems.
    calloutAssertion:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        C2's short-lived RS256 Service Card assertion, verified against the same
        JWKS as id_tokens. `aud` must equal our `client_id`; `exp` is about 60
        seconds after `iat`.
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    Limit:
      name: limit
      in: query
      schema: { type: integer, default: 50, maximum: 200 }
    Offset:
      name: offset
      in: query
      schema: { type: integer, default: 0 }
    Sort:
      name: sort
      in: query
      schema: { type: string }
      description: A sort key; unknown keys fall back to the default.
  responses:
    BadRequest:
      description: Validation failed
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    NotFound:
      description: No such resource
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Conflict:
      description: |
        The request conflicts with current state. Check `code`:
        `stale_version` (reload and retry), `bad_transition` (illegal status
        move), `duplicate`, or `last_admin`.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    Problem:
      type: object
      description: RFC 7807. Switch on `code`, which is stable; `detail` is prose.
      properties:
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        code:
          type: string
          enum:
            [not_found, unauthenticated, no_access, suspended, forbidden,
             stale_version, bad_transition, duplicate, conflict, suppressed,
             invalid_input, invalid_json, rate_limited, internal]
        instance: { type: string }
        requestId: { type: string }
    Page:
      type: object
      properties:
        total: { type: integer }
        limit: { type: integer }
        offset: { type: integer }
        hasMore: { type: boolean }
    RequestStatus:
      type: string
      enum: [new, triaged, assigned, in_progress, waiting_citizen,
             waiting_third_party, resolved, closed, cancelled]
    Priority:
      type: string
      enum: [low, normal, high, urgent, critical]
    ServiceRequest:
      type: object
      properties:
        id: { type: string, format: uuid }
        reference:
          type: string
          example: SR-2026-000123
          description: The human-quotable identifier a citizen reads out.
        contactId: { type: string, format: uuid }
        serviceTypeId: { type: string, format: uuid }
        status: { $ref: "#/components/schemas/RequestStatus" }
        priority: { $ref: "#/components/schemas/Priority" }
        subject: { type: string }
        description: { type: string }
        queueId: { type: string }
        assigneeUserId: { type: string }
        openedAt: { type: string, format: date-time }
        dueAt: { type: string, format: date-time, description: "In business hours, not wall time" }
        slaBreached: { type: boolean }
        version:
          type: integer
          description: Send this back on updates for optimistic concurrency.
    CreateRequest:
      type: object
      required: [contactId]
      properties:
        contactId: { type: string, format: uuid }
        serviceTypeId: { type: string, format: uuid }
        serviceTypeCode:
          type: string
          description: An alternative to serviceTypeId — partners quote codes, not UUIDs.
        subject: { type: string }
        description: { type: string }
        priority: { $ref: "#/components/schemas/Priority" }
        externalRef:
          type: string
          description: The originating system's own identifier, for reconciliation.
        address1: { type: string }
        ward: { type: string }
        formData:
          type: object
          description: Validated against the service type's intake form.
...
    Comment:
      type: object
      properties:
        id: { type: string, format: uuid }
        body: { type: string }
        visibility: { type: string, enum: [internal, citizen] }
        authorName: { type: string }
        createdAt: { type: string, format: date-time }
...